feat: add 'update comment' feature

+++ src/main/java/ilknurdogan/servicefinder/dto/responseDto/ServiceProviderGetDto.java
@@ -22,7 +22,7 @@ public class ServiceProviderGetDto {
 
     private String district;
 
-    private int averageScore;
+    private double averageScore;
 
     private String category;
